requirements.txt
langchain
crewai
langchain_community
langchain_openai
python-dotenv
pre-commit
amadeus
wikipedia
requests
fastapi
uvicorn
crewai[tools]
nltk
streamlit_timeline
